Minor fixes, Moved some code into seperate files for readability
